code_rag_agent/requirements.txt
# Core dependencies
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.24.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
python-dotenv>=1.0.0
pydantic>=2.0.0
rank-bm25>=0.2.0
torch>=2.0.0
transformers>=4.30.0
